backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
earthengine-api==0.1.384
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
googlemaps==4.10.0
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.1
pytest==7.4.3
pytest-cov==4.1.0

